[AUD-1439] Move profile-page store to common
