+++ src/app.tsx
@@ -1,6 +1,6 @@
+import '@audius/stems/dist/stems.css'
 import React from 'react'
 
-import '@audius/stems/dist/stems.css'
 import { ConnectedRouter } from 'connected-react-router'
 import { Provider } from 'react-redux'
 import { Route, Switch } from 'react-router-dom'
