+++ src/common/store/index.ts
@@ -21,6 +21,7 @@ import exploreCollectionsReducer from 'common/store/pages/explore/exploreCollect
 import explorePageReducer from 'common/store/pages/explore/reducer'
 import feed from 'common/store/pages/feed/reducer'
 import { FeedPageState } from 'common/store/pages/feed/types'
+import profileReducer from 'common/store/pages/profile/reducer'
 import tokenDashboardSlice from 'common/store/pages/token-dashboard/slice'
 import track from 'common/store/pages/track/reducer'
 import TrackPageState from 'common/store/pages/track/types'
@@ -91,7 +92,8 @@ export const reducers = {
     exploreCollections: exploreCollectionsReducer,
     tokenDashboard: tokenDashboardSlice.reducer,
     track
-  })
+  }),
+  profile: profileReducer
 }
 
 export const sagas = {
