+++ src/common/store/pages/profile/sagas.js
@@ -22,11 +22,11 @@ import {
 } from 'common/store/cache/users/sagas'
 import { getUser } from 'common/store/cache/users/selectors'
 import { processAndCacheUsers } from 'common/store/cache/users/utils'
+import feedSagas from 'common/store/pages/profile/lineups/feed/sagas.js'
+import tracksSagas from 'common/store/pages/profile/lineups/tracks/sagas.js'
 import { squashNewLines } from 'common/utils/formatUtil'
 import { makeUid, makeKindId } from 'common/utils/uid'
 import * as artistRecommendationsActions from 'components/artist-recommendations/store/slice'
-import feedSagas from 'pages/profile-page/store/lineups/feed/sagas.js'
-import tracksSagas from 'pages/profile-page/store/lineups/tracks/sagas.js'
 import AudiusBackend, { fetchCID } from 'services/AudiusBackend'
 import { setAudiusAccountUser } from 'services/LocalStorage'
 import apiClient from 'services/audius-api-client/AudiusAPIClient'
