+++ src/components/collectibles/components/CollectibleDetails.tsx
@@ -6,11 +6,11 @@ import { useDispatch, useSelector } from 'react-redux'
 import { ReactComponent as IconPlay } from 'assets/img/pbIconPlay.svg'
 import { useModalState } from 'common/hooks/useModalState'
 import { Collectible, CollectibleMediaType } from 'common/models/Collectible'
+import { getProfileUserHandle } from 'common/store/pages/profile/selectors'
 import { setCollectible } from 'common/store/ui/collectible-details/slice'
 import LoadingSpinner from 'components/loading-spinner/LoadingSpinner'
 import PerspectiveCard from 'components/perspective-card/PerspectiveCard'
 import PreloadImage from 'components/preload-image/PreloadImage'
-import { getProfileUserHandle } from 'pages/profile-page/store/selectors'
 import { preload } from 'utils/image'
 
 import { getFrameFromGif } from '../ethCollectibleHelpers'
