+++ src/components/collectibles/components/CollectiblesPage.tsx
@@ -34,6 +34,7 @@ import useInstanceVar from 'common/hooks/useInstanceVar'
 import { useModalState } from 'common/hooks/useModalState'
 import { Collectible, CollectiblesMetadata } from 'common/models/Collectible'
 import { FeatureFlags } from 'common/services/remote-config/feature-flags'
+import { ProfileUser } from 'common/store/pages/profile/types'
 import { getCollectible } from 'common/store/ui/collectible-details/selectors'
 import { setCollectible } from 'common/store/ui/collectible-details/slice'
 import CollectibleDetails from 'components/collectibles/components/CollectibleDetails'
@@ -48,7 +49,6 @@ import { ToastContext } from 'components/toast/ToastContext'
 import { ComponentPlacement, MountPlacement } from 'components/types'
 import UserBadges from 'components/user-badges/UserBadges'
 import { useFlag } from 'hooks/useRemoteConfig'
-import { ProfileUser } from 'pages/profile-page/store/types'
 import { copyToClipboard, getCopyableLink } from 'utils/clipboardUtil'
 import {
   BASE_GA_URL,
