+++ src/components/user-badges/hooks.tsx
@@ -2,7 +2,7 @@ import { useMemo } from 'react'
 
 import { ID } from 'common/models/Identifiers'
 import { getAccountUser } from 'common/store/account/selectors'
-import { getProfileUser } from 'pages/profile-page/store/selectors'
+import { getProfileUser } from 'common/store/pages/profile/selectors'
 import { useSelector } from 'utils/reducer'
 
 import { makeGetTierAndVerifiedForUser } from './utils'
