+++ src/components/user-list/UserList.tsx
@@ -9,9 +9,9 @@ import { ID } from 'common/models/Identifiers'
 import { User } from 'common/models/User'
 import { getUserId } from 'common/store/account/selectors'
 import { getUsers } from 'common/store/cache/users/selectors'
+import { setNotificationSubscription } from 'common/store/pages/profile/actions'
 import * as socialActions from 'common/store/social/users/actions'
 import * as unfollowConfirmationActions from 'components/unfollow-confirmation-modal/store/actions'
-import { setNotificationSubscription } from 'pages/profile-page/store/actions'
 import { AppState } from 'store/types'
 import { isMobile } from 'utils/clientUtil'
 import { profilePage } from 'utils/route'
