+++ src/pages/artist-dashboard-page/store/sagas.js
@@ -4,8 +4,8 @@ import { all, call, put, take, takeEvery } from 'redux-saga/effects'
 
 import { IntKeys } from 'common/services/remote-config'
 import { getAccountUser } from 'common/store/account/selectors'
+import { retrieveUserTracks } from 'common/store/pages/profile/lineups/tracks/retrieveUserTracks'
 import { getBalance } from 'common/store/wallet/slice'
-import { retrieveUserTracks } from 'pages/profile-page/store/lineups/tracks/retrieveUserTracks'
 import AudiusBackend from 'services/AudiusBackend'
 import { remoteConfigInstance } from 'services/remote-config/remote-config-instance'
 import { waitForBackendSetup } from 'store/backend/sagas'
