+++ src/pages/profile-page/components/desktop/ProfilePage.tsx
@@ -12,6 +12,9 @@ import { CoverPhotoSizes, ProfilePictureSizes } from 'common/models/ImageSizes'
 import { LineupState } from 'common/models/Lineup'
 import Status from 'common/models/Status'
 import { User } from 'common/models/User'
+import { feedActions } from 'common/store/pages/profile/lineups/feed/actions'
+import { tracksActions } from 'common/store/pages/profile/lineups/tracks/actions'
+import { ProfileUser, Tabs } from 'common/store/pages/profile/types'
 import Card from 'components/card/desktop/Card'
 import CollectiblesPage from 'components/collectibles/components/CollectiblesPage'
 import CoverPhoto from 'components/cover-photo/CoverPhoto'
@@ -28,9 +31,6 @@ import { badgeTiers } from 'components/user-badges/utils'
 import useTabs, { useTabRecalculator } from 'hooks/useTabs/useTabs'
 import { MIN_COLLECTIBLES_TIER } from 'pages/profile-page/ProfilePageProvider'
 import EmptyTab from 'pages/profile-page/components/EmptyTab'
-import { feedActions } from 'pages/profile-page/store/lineups/feed/actions'
-import { tracksActions } from 'pages/profile-page/store/lineups/tracks/actions'
-import { ProfileUser, Tabs } from 'pages/profile-page/store/types'
 import { make, useRecord } from 'store/analytics/actions'
 import {
   albumPage,
