+++ src/pages/profile-page/components/mobile/ProfilePage.tsx
@@ -13,6 +13,9 @@ import { CoverPhotoSizes, ProfilePictureSizes } from 'common/models/ImageSizes'
 import { LineupState } from 'common/models/Lineup'
 import Status from 'common/models/Status'
 import { User } from 'common/models/User'
+import { feedActions } from 'common/store/pages/profile/lineups/feed/actions'
+import { tracksActions } from 'common/store/pages/profile/lineups/tracks/actions'
+import { Tabs, ProfileUser } from 'common/store/pages/profile/types'
 import Card from 'components/card/mobile/Card'
 import CollectiblesPage from 'components/collectibles/components/CollectiblesPage'
 import { HeaderContext } from 'components/header/mobile/HeaderContextProvider'
@@ -32,9 +35,6 @@ import { badgeTiers } from 'components/user-badges/utils'
 import useAsyncPoll from 'hooks/useAsyncPoll'
 import useTabs from 'hooks/useTabs/useTabs'
 import { MIN_COLLECTIBLES_TIER } from 'pages/profile-page/ProfilePageProvider'
-import { feedActions } from 'pages/profile-page/store/lineups/feed/actions'
-import { tracksActions } from 'pages/profile-page/store/lineups/tracks/actions'
-import { Tabs, ProfileUser } from 'pages/profile-page/store/types'
 import { albumPage, playlistPage, fullProfilePage } from 'utils/route'
 import { withNullGuard } from 'utils/withNullGuard'
