+++ src/store/account/mobileSagas.ts
@@ -4,7 +4,7 @@ import { takeEvery, put, call } from 'redux-saga/effects'
 import { User } from 'common/models/User'
 import * as accountActions from 'common/store/account/reducer'
 import { setNeedsAccountRecovery } from 'common/store/account/reducer'
-import { updateProfileAsync } from 'pages/profile-page/store/sagas'
+import { updateProfileAsync } from 'common/store/pages/profile/sagas'
 import AudiusBackend from 'services/AudiusBackend'
 import { FetchAccountFailed } from 'services/native-mobile-interface/lifecycle'
 import { ReloadMessage } from 'services/native-mobile-interface/linking'
