+++ src/store/playlist-library/sagas.ts
@@ -25,8 +25,8 @@ import {
   getPlaylistLibrary
 } from 'common/store/account/selectors'
 import * as cacheActions from 'common/store/cache/actions'
+import { updateProfileAsync } from 'common/store/pages/profile/sagas'
 import { makeKindId } from 'common/utils/uid'
-import { updateProfileAsync } from 'pages/profile-page/store/sagas'
 import { make } from 'store/analytics/actions'
 import { waitForBackendSetup } from 'store/backend/sagas'
 import { getResult } from 'store/confirmer/selectors'
