+++ src/store/reducers.js
@@ -2,6 +2,7 @@ import { connectRouter } from 'connected-react-router'
 import { combineReducers } from 'redux'
 
 import { reducers as clientStoreReducers } from 'common/store'
+import profile from 'common/store/pages/profile/reducer'
 import remoteConfig from 'common/store/remote-config/slice'
 import artistRecommendations from 'components/artist-recommendations/store/slice'
 import changePassword from 'components/change-password/store/slice'
@@ -24,7 +25,6 @@ import followers from 'pages/followers-page/store/reducers'
 import following from 'pages/following-page/store/reducers'
 import history from 'pages/history-page/store/reducer'
 import notificationUsers from 'pages/notification-users-page/store/reducers'
-import profile from 'pages/profile-page/store/reducer'
 import remixes from 'pages/remixes-page/store/slice'
 import reposts from 'pages/reposts-page/store/reducers'
 import saved from 'pages/saved-page/store/reducer'
