+++ src/store/sagas.js
@@ -7,6 +7,7 @@ import usersSagas from 'common/store/cache/users/sagas'
 import errorSagas from 'common/store/errors/sagas'
 import exploreCollectionsPageSagas from 'common/store/pages/explore/exploreCollections/sagas'
 import explorePageSagas from 'common/store/pages/explore/sagas'
+import profileSagas from 'common/store/pages/profile/sagas'
 import remoteConfigSagas from 'common/store/remote-config/sagas'
 import shareModalSagas from 'common/store/ui/share-modal/sagas'
 import toastSagas from 'common/store/ui/toast/sagas'
@@ -31,7 +32,6 @@ import followersPageSagas from 'pages/followers-page/store/sagas'
 import followingPageSagas from 'pages/following-page/store/sagas'
 import historySagas from 'pages/history-page/store/sagas'
 import notificationUsersPageSagas from 'pages/notification-users-page/store/sagas'
-import profileSagas from 'pages/profile-page/store/sagas'
 import remixesSagas from 'pages/remixes-page/store/sagas'
 import repostPageSagas from 'pages/reposts-page/store/sagas'
 import savedSagas from 'pages/saved-page/store/sagas'
