+++ src/utils/route/userRouteParser.ts
@@ -1,7 +1,7 @@
 import { matchPath } from 'react-router-dom'
 
 import { ID } from 'common/models/Identifiers'
-import { TabRoute } from 'pages/profile-page/store/types'
+import { TabRoute } from 'common/store/pages/profile/types'
 import { USER_ID_PAGE, PROFILE_PAGE, staticRoutes } from 'utils/route'
 
 import { decodeHashId } from './hashIds'
